:recycle: Refactor RegisterProductController to return a header location to DetailProductController to show all data about the product that was created
