+++ src/main/java/com/mercadolivre/apimlv2/usecases/detailsproduct/DetailsProductController.java
@@ -1,7 +1,6 @@
 package com.mercadolivre.apimlv2.usecases.detailsproduct;
 
 import com.mercadolivre.apimlv2.domain.Product;
-import com.mercadolivre.apimlv2.usecases.registerproduct.RegisterProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,11 +16,11 @@ public class DetailsProductController {
     private EntityManager manager;
 
     @GetMapping("/products/{productId}")
-    public ResponseEntity<RegisterProductResponse> detailsProduct(@PathVariable("productId") Long productId) {
+    public ResponseEntity<DetailProductResponse> detailsProduct(@PathVariable("productId") Long productId) {
         Product product = manager.find(Product.class, productId);
         if (product == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(new RegisterProductResponse(product));
+        return ResponseEntity.ok(new DetailProductResponse(product));
     }
 }
