+++ src/main/java/com/mercadolivre/apimlv2/usecases/registerproduct/RegisterProductController.java
@@ -3,17 +3,18 @@ package com.mercadolivre.apimlv2.usecases.registerproduct;
 import com.mercadolivre.apimlv2.domain.Product;
 import com.mercadolivre.apimlv2.domain.User;
 import com.mercadolivre.apimlv2.security.LoggedUser;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+import java.net.URI;
 
 @RestController
 public class RegisterProductController {
@@ -23,10 +24,12 @@ public class RegisterProductController {
 
     @PostMapping("/products")
     @Transactional
-    public ResponseEntity<RegisterProductResponse> registerProduct(@RequestBody @Valid RegisterProductRequest request, @AuthenticationPrincipal LoggedUser loggedUser) {
+    public ResponseEntity<Void> registerProduct(@RequestBody @Valid RegisterProductRequest request, @AuthenticationPrincipal LoggedUser loggedUser, UriComponentsBuilder uriComponentsBuilder) {
         User productOwner = loggedUser.get();
         Product newProduct = request.toModel(productOwner, manager);
         manager.persist(newProduct);
-        return ResponseEntity.status(HttpStatus.CREATED).body(new RegisterProductResponse(newProduct));
+
+        URI url = uriComponentsBuilder.path("/products/{id}").buildAndExpand(newProduct.getId()).toUri();
+        return ResponseEntity.created(url).build();
     }
 }
